Switch over to gotelnats module
