+++ go.mod
@@ -4,22 +4,22 @@ go 1.18
 
 require (
 	github.com/cyverse-de/configurate v0.0.0-20220113221928-13d34aae3f0f
+	github.com/cyverse-de/go-mod/gotelnats v0.0.4
 	github.com/cyverse-de/go-mod/logging v0.0.1
-	github.com/cyverse-de/go-mod/protobufjson v0.0.1
-	github.com/cyverse-de/p/go/svcerror v0.0.2
-	github.com/cyverse-de/p/go/user v0.0.5
+	github.com/cyverse-de/go-mod/protobufjson v0.0.2
+	github.com/cyverse-de/p/go/svcerror v0.0.3
+	github.com/cyverse-de/p/go/user v0.0.7
 	github.com/doug-martin/goqu/v9 v9.18.0
 	github.com/jmoiron/sqlx v1.3.4
 	github.com/lib/pq v1.10.1
-	github.com/nats-io/nats.go v1.13.1-0.20220308171302-2f2f6968e98d
+	github.com/nats-io/nats.go v1.14.0
 	github.com/sirupsen/logrus v1.8.1
 	github.com/spf13/viper v1.10.1
-	go.opentelemetry.io/otel v1.6.3
-	go.opentelemetry.io/otel/trace v1.6.3
 	google.golang.org/protobuf v1.28.0
 )
 
 require (
+	github.com/cyverse-de/p/go/header v0.0.0-20220425195853-720402e94a27 // indirect
 	github.com/fsnotify/fsnotify v1.5.1 // indirect
 	github.com/go-logr/logr v1.2.3 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
@@ -35,6 +35,8 @@ require (
 	github.com/spf13/jwalterweatherman v1.1.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
 	github.com/subosito/gotenv v1.2.0 // indirect
+	go.opentelemetry.io/otel v1.6.3 // indirect
+	go.opentelemetry.io/otel/trace v1.6.3 // indirect
 	golang.org/x/crypto v0.0.0-20220112180741-5e0467b6c7ce // indirect
 	golang.org/x/sys v0.0.0-20220403205710-6acee93ad0eb // indirect
 	golang.org/x/text v0.3.7 // indirect
